+++ src/delfino/constants.py
@@ -1,8 +1,9 @@
 from enum import Enum
+from pathlib import Path
 
 ENTRY_POINT: str = "delfino"
 PYPROJECT_TOML_FILENAME = "pyproject.toml"
-COMMANDS_DIRECTORY_NAME = "commands"
+DEFAULT_LOCAL_COMMANDS_DIRECTORY = Path("commands")
 
 
 class PackageManager(Enum):